refactor：storybook：component：LoginMainParts：関数切り出し
