+++ src/stories/components/organisms/Main/LogsMainParts/index.tsx
@@ -1,12 +1,10 @@
 import ReceiptLongIcon from '@mui/icons-material/ReceiptLong';
-import { FC, useState } from 'react';
+import { FC } from 'react';
+
+import { useLogsMainFunctions } from './hooks';
 
 import theme from '@/stories/common/theme';
-import {
-  AddressListType,
-  MessageContentsType,
-  MessageType,
-} from '@/stories/common/types';
+import { AddressListType } from '@/stories/common/types';
 import { Box } from '@/stories/components/atoms/Box/Basic';
 import { IconButton } from '@/stories/components/atoms/IconButton/Base';
 import { Paper } from '@/stories/components/atoms/Paper/Base';
@@ -19,40 +17,9 @@ export interface LogsListProp {
   personId: string;
 }
 
-export const LogsMainParts: FC<LogsListProp> = ({ data, personId }: LogsListProp) => {
-  const [displayLogFlg, setDisplayLogFlg] = useState(false);
-  const [messageLog, setMessageLog] = useState<MessageContentsType[]>([]);
-  const orgMessageLog = data[personId] && data[personId].message;
-  const logExistFlg = orgMessageLog
-    ? (orgMessageLog.from && orgMessageLog.from.length) ||
-      (orgMessageLog.to && orgMessageLog.to.length)
-    : false;
-  const handleClickDisplayLogs = () => {
-    setDisplayLogFlg(true);
-    sortMessage();
-  };
-
-  const sortMessage = () => {
-    setMessageLog([]);
-    const unionMessage: MessageContentsType[] = sortedUnionMessage(orgMessageLog);
-    setMessageLog(unionMessage);
-  };
-
-  const sortedUnionMessage = (userMessage: MessageType): MessageContentsType[] => {
-    const unionMessage: MessageContentsType[] = [];
-    const fieldary: ('to' | 'from')[] = ['to', 'from'];
-    fieldary.forEach((key) => {
-      if (userMessage && userMessage[key]) {
-        userMessage[key].forEach((item: MessageContentsType) => {
-          unionMessage.push(item);
-        });
-      }
-    });
-    unionMessage.sort((x, y) => {
-      return x.sortTime - y.sortTime;
-    });
-    return unionMessage;
-  };
+export const LogsMainParts: FC<LogsListProp> = (props: LogsListProp) => {
+  const { logExistFlg, handleClickDisplayLogs, displayLogFlg, messageLog } =
+    useLogsMainFunctions(props);
 
   return (
     <Paper sx={{ margin: theme.spacing(0.5), padding: theme.spacing(2) }}>
